feat:  FightProp rewrite, avatar/weapon/inventory overhaul, entity decoupling & proto normalization

+++ GameServer/Server/Packet/Send/Avatar/PacketAvatarDieAnimationEndRsp.cs
@@ -10,7 +10,7 @@ public class PacketAvatarDieAnimationEndRsp : BasePacket
         var proto = new AvatarDieAnimationEndRsp
         {
             DieGuid = dieGuid,
-            Retcode = (uint)retcode
+            Retcode = retcode
         };
 
         SetData(proto);
